refactor (tests): move backend functions to ... backends.py
